+++ packages/insomnia-app/app/models/helpers/base-space.ts
@@ -1,10 +1,10 @@
 import { getAppName } from '../../common/constants';
-import { Space, BASE_SPACE_ID } from '../space';
+import { BASE_SPACE_ID, Space } from '../space';
 
 export type SpaceSubset = Pick<Space, '_id' | 'name' | 'remoteId'>;
 
 // TODO: this is a holdover until we have a default space always existing
-export const defaultSpace: SpaceSubset = {
+export const baseSpace: SpaceSubset = {
   _id: BASE_SPACE_ID,
   name: getAppName(),
   remoteId: null,
